% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertText.R
\name{insertEqualSignAddin}
\alias{insertEqualSignAddin}
\title{Insert " = ".}
\usage{
insertEqualSignAddin()
}
\description{
Call this function as an addin to insert \code{" = "} at the cursor position.
}
