% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceText.R
\name{replaceBackWithForwardSlashAddin}
\alias{replaceBackWithForwardSlashAddin}
\title{Replace "\" with "/".}
\usage{
replaceBackWithForwardSlashAddin()
}
\description{
Call this function as an addin to replace backslashes \code{\\} with forward
slashes \code{/} in the selected text.
}
